+++ TFY4235_numfys/2/src/physics.c
@@ -1,7 +1,9 @@
 #include "physics.h"
+#include "utils.h"
 
 #include <stdio.h>
 #include <math.h>
+
 
 int f_t(double t){
     if (constants->TAU == 0.0) return 0;
@@ -9,11 +11,11 @@ int f_t(double t){
 }
 
 double U_r(double x){
-    if (0 <= fmod(x, 1) && fmod(x, 1) < constants->ALPHA){
-        return x / constants->ALPHA;
+    if (is_in_mod_range(x, 0, constants->ALPHA * constants->L, constants->L)){
+        return x / (constants->ALPHA * constants->L);
     }
-    else if (constants->ALPHA <= fmod(x, 1) && fmod(x, 1) < 1){
-        return (1 - x) / (1 - constants->ALPHA);
+    else if (is_in_mod_range(x, constants->ALPHA * constants->L, constants->L, constants->L)){
+        return (constants->L - x) / (constants->L * (1 - constants->ALPHA));
     }
     // sanity check
     else return 0.0;
